Moved constraint extraction functions from MilpBipartite to cons.cpp
